# backend/app/api/routes/health.py
from fastapi import APIRouter, Depends, Query, status
from fastapi.responses import JSONResponse
from sqlalchemy.exc import IntegrityError
from sqlalchemy.orm import Session

from app.api.deps.auth import AuthenticatedContext, require_authenticated_context
from app.core.config import get_settings
from app.contracts.common import DataEnvelope, ErrorBody, ErrorEnvelope
from app.contracts.system import (
    DemoBootstrapSummaryData,
    HealthResponse,
    IncidentTimelineData,
    IncidentTaskDiagnosticReplayData,
    IncidentTaskDiagnosticReplayRequest,
    OperatorDiagnosticsData,
    OperatorViewBackfillData,
    OperatorViewBackfillRequest,
    PilotControlData,
    PilotControlMutationRequest,
    PilotSummaryData,
    ShiftBundleExportRecordData,
    ShiftBundleExportRecordRequest,
    SystemReadinessData,
)
from app.db.session import get_db_session
from app.models import PilotControl
from app.services.demo_state import bootstrap_demo_state
from app.services.pilot_control import (
    PilotControlMutationResult,
    PilotControlTransitionError,
    PilotControlValidationError,
    get_or_create_pilot_control,
    mutate_pilot_control,
)
from app.services.pilot_incidents import (
    IncidentRecoveryTaskNotFoundError,
    build_incident_timeline,
    build_operator_diagnostics,
    build_operator_view_backfill,
    build_task_diagnostic_replay,
    record_shift_bundle_export,
)
from app.services.pilot_summary import build_pilot_summary
from app.services.system_readiness import build_system_readiness

router = APIRouter(tags=["system"])


def _unauthorized() -> JSONResponse:
    return JSONResponse(
        status_code=status.HTTP_401_UNAUTHORIZED,
        content=ErrorEnvelope(
            error=ErrorBody(code="unauthorized", message="Unauthorized", details=[])
        ).model_dump(),
    )


def _not_found() -> JSONResponse:
    return JSONResponse(
        status_code=status.HTTP_404_NOT_FOUND,
        content=ErrorEnvelope(
            error=ErrorBody(code="shop_not_found", message="Shop not found", details=[])
        ).model_dump(),
    )


def _error_response(status_code: int, code: str, message: str) -> JSONResponse:
    return JSONResponse(
        status_code=status_code,
        content=ErrorEnvelope(
            error=ErrorBody(code=code, message=message, details=[])
        ).model_dump(),
    )


def _to_pilot_control_data(
    pilot_control: PilotControl,
    *,
    previous_cutover_mode: str | None = None,
    transition_audit_log_id: str | None = None,
) -> PilotControlData:
    data = PilotControlData.model_validate(pilot_control, from_attributes=True)
    if previous_cutover_mode is None and transition_audit_log_id is None:
        return data
    return data.model_copy(
        update={
            "previous_cutover_mode": previous_cutover_mode,
            "transition_audit_log_id": transition_audit_log_id,
        }
    )


def _ensure_pilot_control_with_retry(
    db_session: Session,
    *,
    shop_id: str,
    trial_provider_profile: str,
) -> PilotControl:
    pilot_control, changed = get_or_create_pilot_control(
        db_session,
        shop_id=shop_id,
        trial_provider_profile=trial_provider_profile,
    )
    if not changed:
        return pilot_control

    try:
        db_session.commit()
        return pilot_control
    except IntegrityError:
        db_session.rollback()
        pilot_control, changed = get_or_create_pilot_control(
            db_session,
            shop_id=shop_id,
            trial_provider_profile=trial_provider_profile,
        )
        if changed:
            db_session.commit()
        return pilot_control


def _mutate_pilot_control_with_retry(
    db_session: Session,
    *,
    shop_id: str,
    actor_id: str,
    trial_provider_profile: str,
    payload: PilotControlMutationRequest,
) -> PilotControlMutationResult:
    try:
        mutation_result = mutate_pilot_control(
            db_session,
            shop_id=shop_id,
            actor_id=actor_id,
            trial_provider_profile=trial_provider_profile,
            cutover_mode=payload.cutover_mode,
            approved_calibration_artifact_id=payload.approved_calibration_artifact_id,
            approved_calibration_report_path=payload.approved_calibration_report_path,
            last_preflight_status=payload.last_preflight_status,
            notes=payload.notes,
        )
        if mutation_result.changed:
            db_session.commit()
        return mutation_result
    except IntegrityError:
        db_session.rollback()
        mutation_result = mutate_pilot_control(
            db_session,
            shop_id=shop_id,
            actor_id=actor_id,
            trial_provider_profile=trial_provider_profile,
            cutover_mode=payload.cutover_mode,
            approved_calibration_artifact_id=payload.approved_calibration_artifact_id,
            approved_calibration_report_path=payload.approved_calibration_report_path,
            last_preflight_status=payload.last_preflight_status,
            notes=payload.notes,
        )
        if mutation_result.changed:
            db_session.commit()
        return mutation_result


@router.get("/health", response_model=HealthResponse)
def health() -> HealthResponse:
    return HealthResponse(status="ok")


@router.get(
    "/api/v1/system/readiness",
    response_model=DataEnvelope[SystemReadinessData],
    responses={status.HTTP_401_UNAUTHORIZED: {"model": ErrorEnvelope, "description": "Unauthorized"}},
)
def get_system_readiness(
    auth: AuthenticatedContext = Depends(require_authenticated_context),
    db_session: Session = Depends(get_db_session),
) -> DataEnvelope[SystemReadinessData]:
    settings = get_settings()
    pilot_control = _ensure_pilot_control_with_retry(
        db_session,
        shop_id=auth.shop_id,
        trial_provider_profile=settings.trial_provider_profile.strip(),
    )
    readiness_data: SystemReadinessData = build_system_readiness(settings).model_copy(
        update={
            "trial_provider_profile": pilot_control.trial_provider_profile,
            "approved_calibration_artifact_id": pilot_control.approved_calibration_artifact_id,
            "cutover_mode": pilot_control.cutover_mode,
        }
    )
    return DataEnvelope(data=readiness_data)


@router.get(
    "/api/v1/system/pilot-summary",
    response_model=DataEnvelope[PilotSummaryData],
    responses={status.HTTP_401_UNAUTHORIZED: {"model": ErrorEnvelope, "description": "Unauthorized"}},
)
def get_pilot_summary(
    auth: AuthenticatedContext = Depends(require_authenticated_context),
    db_session: Session = Depends(get_db_session),
    hours: int = Query(default=24, ge=1, le=168),
) -> DataEnvelope[PilotSummaryData]:
    settings = get_settings()
    return DataEnvelope(
        data=build_pilot_summary(
            db_session,
            shop_id=auth.shop_id,
            hours=hours,
            trial_provider_profile=settings.trial_provider_profile.strip(),
        )
    )


@router.get(
    "/api/v1/system/incident-timeline",
    response_model=DataEnvelope[IncidentTimelineData],
    responses={status.HTTP_401_UNAUTHORIZED: {"model": ErrorEnvelope, "description": "Unauthorized"}},
)
def get_incident_timeline(
    auth: AuthenticatedContext = Depends(require_authenticated_context),
    db_session: Session = Depends(get_db_session),
    hours: int = Query(default=24, ge=1, le=168),
    limit: int = Query(default=20, ge=1, le=100),
) -> DataEnvelope[IncidentTimelineData]:
    settings = get_settings()
    return DataEnvelope(
        data=build_incident_timeline(
            db_session,
            shop_id=auth.shop_id,
            hours=hours,
            limit=limit,
            trial_provider_profile=settings.trial_provider_profile.strip(),
        )
    )


@router.post(
    "/api/v1/system/incident-replays/task-diagnostic",
    response_model=DataEnvelope[IncidentTaskDiagnosticReplayData],
    responses={
        status.HTTP_401_UNAUTHORIZED: {"model": ErrorEnvelope, "description": "Unauthorized"},
        status.HTTP_404_NOT_FOUND: {"model": ErrorEnvelope, "description": "Task run not found"},
    },
)
def post_task_diagnostic_replay(
    payload: IncidentTaskDiagnosticReplayRequest,
    auth: AuthenticatedContext = Depends(require_authenticated_context),
    db_session: Session = Depends(get_db_session),
) -> DataEnvelope[IncidentTaskDiagnosticReplayData] | JSONResponse:
    settings = get_settings()
    try:
        replay = build_task_diagnostic_replay(
            db_session,
            shop_id=auth.shop_id,
            actor_id=auth.actor_id,
            task_run_id=payload.task_run_id,
            idempotency_key=payload.idempotency_key,
            trial_provider_profile=settings.trial_provider_profile.strip(),
        )
        db_session.commit()
    except IncidentRecoveryTaskNotFoundError:
        db_session.rollback()
        return _error_response(status.HTTP_404_NOT_FOUND, "task_run_not_found", "Task run not found")
    except Exception:
        db_session.rollback()
        raise
    return DataEnvelope(data=replay)


@router.post(
    "/api/v1/system/operator-view-backfills",
    response_model=DataEnvelope[OperatorViewBackfillData],
    responses={status.HTTP_401_UNAUTHORIZED: {"model": ErrorEnvelope, "description": "Unauthorized"}},
)
def post_operator_view_backfill(
    payload: OperatorViewBackfillRequest,
    auth: AuthenticatedContext = Depends(require_authenticated_context),
    db_session: Session = Depends(get_db_session),
) -> DataEnvelope[OperatorViewBackfillData]:
    settings = get_settings()
    backfill = build_operator_view_backfill(
        db_session,
        shop_id=auth.shop_id,
        actor_id=auth.actor_id,
        view=payload.view,
        hours=payload.hours,
        limit=payload.limit,
        idempotency_key=payload.idempotency_key,
        trial_provider_profile=settings.trial_provider_profile.strip(),
    )
    db_session.commit()
    return DataEnvelope(data=backfill)


@router.get(
    "/api/v1/system/operator-diagnostics",
    response_model=DataEnvelope[OperatorDiagnosticsData],
    responses={status.HTTP_401_UNAUTHORIZED: {"model": ErrorEnvelope, "description": "Unauthorized"}},
)
def get_operator_diagnostics(
    auth: AuthenticatedContext = Depends(require_authenticated_context),
    db_session: Session = Depends(get_db_session),
    hours: int = Query(default=24, ge=1, le=168),
    limit: int = Query(default=20, ge=1, le=100),
) -> DataEnvelope[OperatorDiagnosticsData]:
    settings = get_settings()
    return DataEnvelope(
        data=build_operator_diagnostics(
            db_session,
            shop_id=auth.shop_id,
            hours=hours,
            limit=limit,
            trial_provider_profile=settings.trial_provider_profile.strip(),
        )
    )


@router.post(
    "/api/v1/system/shift-bundle-exports",
    response_model=DataEnvelope[ShiftBundleExportRecordData],
    responses={status.HTTP_401_UNAUTHORIZED: {"model": ErrorEnvelope, "description": "Unauthorized"}},
)
def post_shift_bundle_export_record(
    payload: ShiftBundleExportRecordRequest,
    auth: AuthenticatedContext = Depends(require_authenticated_context),
    db_session: Session = Depends(get_db_session),
) -> DataEnvelope[ShiftBundleExportRecordData]:
    settings = get_settings()
    result = record_shift_bundle_export(
        db_session,
        shop_id=auth.shop_id,
        actor_id=auth.actor_id,
        bundle_id=payload.bundle_id,
        manifest_path=payload.manifest_path,
        overall_status=payload.overall_status,
        degraded_reasons=list(payload.degraded_reasons),
        cutover_mode=payload.cutover_mode,
        hours=payload.hours,
        trial_provider_profile=settings.trial_provider_profile.strip(),
    )
    db_session.commit()
    return DataEnvelope(data=result)


@router.post("/api/v1/system/demo/bootstrap", response_model=DataEnvelope[DemoBootstrapSummaryData])
def post_demo_bootstrap(
    auth: AuthenticatedContext = Depends(require_authenticated_context),
    db_session: Session = Depends(get_db_session),
) -> DataEnvelope[DemoBootstrapSummaryData] | JSONResponse:
    settings = get_settings()
    if auth.shop_id != settings.default_shop_id:
        return _not_found()
    summary = bootstrap_demo_state(db_session)
    return DataEnvelope(
        data=DemoBootstrapSummaryData(
            **summary.to_dict(),
        )
    )


@router.get(
    "/api/v1/system/pilot-control",
    response_model=DataEnvelope[PilotControlData],
    responses={status.HTTP_401_UNAUTHORIZED: {"model": ErrorEnvelope, "description": "Unauthorized"}},
)
def get_pilot_control(
    auth: AuthenticatedContext = Depends(require_authenticated_context),
    db_session: Session = Depends(get_db_session),
) -> DataEnvelope[PilotControlData]:
    settings = get_settings()
    pilot_control = _ensure_pilot_control_with_retry(
        db_session,
        shop_id=auth.shop_id,
        trial_provider_profile=settings.trial_provider_profile.strip(),
    )
    return DataEnvelope(data=_to_pilot_control_data(pilot_control))


@router.post(
    "/api/v1/system/pilot-control",
    response_model=DataEnvelope[PilotControlData],
    responses={
        status.HTTP_401_UNAUTHORIZED: {"model": ErrorEnvelope, "description": "Unauthorized"},
        status.HTTP_409_CONFLICT: {"model": ErrorEnvelope, "description": "Invalid cutover transition"},
        status.HTTP_422_UNPROCESSABLE_ENTITY: {"model": ErrorEnvelope, "description": "Validation error"},
    },
)
def post_pilot_control(
    payload: PilotControlMutationRequest,
    auth: AuthenticatedContext = Depends(require_authenticated_context),
    db_session: Session = Depends(get_db_session),
) -> DataEnvelope[PilotControlData] | JSONResponse:
    if (
        payload.cutover_mode is None
        and payload.approved_calibration_artifact_id is None
        and payload.approved_calibration_report_path is None
        and payload.last_preflight_status is None
        and payload.notes is None
    ):
        return _error_response(
            status.HTTP_422_UNPROCESSABLE_ENTITY,
            "validation_error",
            "At least one mutable pilot-control field is required.",
        )

    settings = get_settings()
    try:
        mutation_result = _mutate_pilot_control_with_retry(
            db_session,
            shop_id=auth.shop_id,
            actor_id=auth.actor_id,
            trial_provider_profile=settings.trial_provider_profile.strip(),
            payload=payload,
        )
    except PilotControlValidationError:
        db_session.rollback()
        return _error_response(
            status.HTTP_422_UNPROCESSABLE_ENTITY,
            "invalid_cutover_mode",
            "cutover_mode is invalid",
        )
    except PilotControlTransitionError:
        db_session.rollback()
        return _error_response(
            status.HTTP_409_CONFLICT,
            "invalid_cutover_mode_transition",
            "cutover_mode transition is not allowed",
        )
    except Exception:
        db_session.rollback()
        raise

    return DataEnvelope(
        data=_to_pilot_control_data(
            mutation_result.pilot_control,
            previous_cutover_mode=mutation_result.previous_cutover_mode,
            transition_audit_log_id=mutation_result.transition_audit_log_id,
        )
    )